Updated CMakeLists.txt  - Added more sub-folders in the source-structure  - Added an option for MSVC to use /bigobj with the Compiler as PrismParser.cpp bloats the object instance count  - Edited CUDD Link Targets for MSVC Edited SymbolicModelAdapter.h, added an alternative implementation for log2 (NOT part of C90, not of Cxx!) Edited Program.cpp, promoted vars from int to uint to conquer warnings related to loss of precision Likewise in DeterministicSparseTransitionParser.cpp, IntegerConstantExpression.h
Edited storm.cpp, reimplemented Usage-Query for non-Unix platforms.
Edited CuddUtility.h, added an include for int Type definitions as they do not fall from the sky
Edited ErrorHandling.h. reimplemented ErrorHandling for non-Unix platforms. Backtraces can not yet be provided.

+++ src/ir/expressions/IntegerConstantExpression.h
@@ -62,7 +62,7 @@ public:
 		return defined;
 	}
 
-	int getValue() {
+	int_fast64_t getValue() {
 		return value;
 	}
